+++ tests/rest/BaseApiModuleCept.php
@@ -1,0 +1,10 @@
+<?php
+$I = new RestTester($scenario);
+
+$I->wantToTest('default api module');
+
+$I->sendGET('/hello');
+$I->seeResponseContainsJson([
+    'message' => 'Hello from Default Api Module: Hello World',
+    'baseEndpointPath' => '/hello'
+]);