+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.0 - 2016-11-XX
+### Changed
+- All not modules endpoints associated with default module in Rest Server.  
+
 ## 0.3.0 - 2016-11-25
 ### Added
 - Api Modules.